# Backend/src/main/java/com/expenseflow/dto/category/CategoryAnalyticDto.java
package com.expenseflow.dto.category;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class CategoryAnalyticDto {
    private CategoryDto category;

    private Double amount;

    private Long numberOfRecords;
}